+++ common/src/main/java/net/caffeinemc/mods/lithium/mixin/world/block_entity_ticking/sleeping/sculk_sensor_shrieker/package-info.java
@@ -1,4 +1,4 @@
 @MixinConfigOption(description = "BlockEntity sleeping for inactive sculk sensors")
-package net.caffeinemc.mods.lithium.mixin.world.block_entity_ticking.sleeping.sculk_sensor;
+package net.caffeinemc.mods.lithium.mixin.world.block_entity_ticking.sleeping.sculk_sensor_shrieker;
 
 import net.caffeinemc.gradle.MixinConfigOption;
