+++ common/src/main/resources/lithium.mixins.json
@@ -225,10 +225,10 @@
     "world.block_entity_ticking.sleeping.crafter.CrafterBlockEntityMixin",
     "world.block_entity_ticking.sleeping.furnace.AbstractFurnaceBlockEntityMixin",
     "world.block_entity_ticking.sleeping.hopper.HopperBlockEntityMixin",
-    "world.block_entity_ticking.sleeping.sculk_sensor.CalibratedSculkSensorBlockMixin",
-    "world.block_entity_ticking.sleeping.sculk_sensor.SculkSensorBlockEntityMixin",
-    "world.block_entity_ticking.sleeping.sculk_sensor.SculkSensorBlockMixin",
-    "world.block_entity_ticking.sleeping.sculk_sensor.VibrationSystemListenerMixin",
+    "world.block_entity_ticking.sleeping.sculk_sensor_shrieker.CalibratedSculkSensorBlockMixin",
+    "world.block_entity_ticking.sleeping.sculk_sensor_shrieker.SculkSensorBlockEntityMixin",
+    "world.block_entity_ticking.sleeping.sculk_sensor_shrieker.SculkSensorBlockMixin",
+    "world.block_entity_ticking.sleeping.sculk_sensor_shrieker.VibrationSystemListenerMixin",
     "world.block_entity_ticking.sleeping.shulker_box.ShulkerBoxBlockEntityMixin",
     "world.block_entity_ticking.world_border.DirectBlockEntityTickInvokerMixin",
     "world.chunk_access.ChunkHolderMixin",
